Break tip panels in half along MCL; leave off in code
Unfortunately, splitting the tip panels in two made the results even
worse than before (large negative doublet strengths, weird Vy splotches
on surface), so we'll leave this option off in the code. It's possible
there is a bug somewhere causing the issues, or it's just not a good
idea to put panels on the tips.