Switch to flat tip caps with BC relaxation for LE tris
The rounded caps were supposed to give smoother variation of velocity
across the tip, but that didn't work. The code is left in there in case
I want to investigate it again later. The BC relaxation is accomplished
by moving the collocation point off the LE tris by 2% of the tip chord,
which seems to work pretty well for getting rid of the high reversed
velocities there. May want to revisit it in the future, though.