Use gage pressure to compute forces and moments
Makes no difference for forces, but can make a difference for pitching
moment when the wing(s) has dihedral. I need to think more about why
this is the case and make sure it is correct. FWIW, AVL results for
the wonky_singlewing case match up with the pitching moment calculated
from relative pressure loads.