Switch to midpoints between trailing legs for Trefftz velocity calc
Seems to be wrong to calculate it at the legs, since that leg then
has no contribution. This causes, for example, the induced velocity
to be in the upward direction at the tip, which is the wrong limiting
behavior. However, between the wingtips, I would have thought the
right limiting behavior would be reached, but there are still some
differences. Warrants further attention.