Switch to P-G compressibility correction and apply to Trefftz plane
Unfortunately, the nonlinear corrections like Karman-Tsien are a
function of the surface pressure distribution, so they cannot be
applied to the integrated quantities in the Trefftz plane. Prandtl-
Glauert is a constant factor, which gets around that issue. Despite
P-G being less accurate, I feel its best to be consistent and use a
single correction for both the surface pressures and the integrated
quantities.